moved services into separate shello/services directory
