+++ src/main/java/hello/AuthControllerAdvice.java
@@ -1,5 +1,6 @@
 package hello;
 
+import hello.services.MembershipService;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
