+++ src/main/java/hello/controllers/EarthquakesController.java
@@ -1,5 +1,5 @@
 package hello.controllers;
-import hello.EarthquakeQueryService;
+import hello.services.EarthquakeQueryService;
 import hello.EqSearch;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
