+++ src/main/java/hello/controllers/HomeController.java
@@ -1,6 +1,6 @@
 package hello.controllers;
 
-import hello.EarthquakeQueryService;
+import hello.services.EarthquakeQueryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
