+++ src/main/java/hello/services/GithubOrgMembershipService.java
@@ -1,4 +1,4 @@
-package hello;
+package hello.services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
