+++ src/main/java/hello/services/MembershipService.java
@@ -1,4 +1,4 @@
-package hello;
+package hello.services;
 
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
